refactor(data-integrity): extract data integrity loading logic
